osutil/disks: return maj/min from mount point if DEVTYPE on mount src is disk
This allows disks.DiskFromMountPoint to be slightly more generic, but will do
the wrong thing for encrypted disks currently, so we need to fix that shortly.

Also add some tests that we don't support any DEVTYPE's other than disk, and
that we do in fact support disk. This also adds coverage of the udevadm call we
are doing with MockCommand.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>