osutil/disks: add HasPartitions, allow non-partition backed decrypted devices
Add a function HasPartitions which dictates whether the disk that corresponds to
the mountpoint's source has partitions or not.

This allows us to have a decrypted device which does not have a backing physical
partition, but instead some kind of device volume or other thing to show up, but
note that we do not expect such a device "in the wild" on uc20, because there
the decrypted devices we care about should be backed by physical partitions.

Also add a unit test around this and add additional checks around this property
to other unit tests.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>