tests: support non-option commands to session-tool
Originally session-tool started with dash-dash options for triggering
certain behavior, most prominently --prepare and --restore. This patch
adds equivalent non-dash options for everything, in addition to the
brand new "exec" command that replaces the undecorated "session-tool
foo" invocation, to run "foo" inside a session.

Note that almost no test changes are peformed, as this ought to be
entirely backwards compatible. Subsequent patches will move the existing
test catalogue over and at some point we will sunset the old names.
The only exception is the test for session-tool, which nows shows the
new syntax in use.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>