interfaces/block_devices: add NVMe subsystem devices, support multipath paths
NVMe devices commonly show up both as /dev/nvme0 and as /dev/nvme0n1, with the
former path actually being an NVMe subsystem device and the latter being a block
subsystem device. This means that the current combination of AppArmor and
devices cgroup UDev rules meant that block-devices could never be fully used for
NVMe devices.

We should support both the /dev/nvme0 and /dev/nvme0n1 paths with block-devices.

Also, we should use the more precise pattern that was used in the k8s-support 
interface more recently, which limits the namespace to 63 and also disallows the
erroneous path /dev/nvme00

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>