interfaces/builtin/desktop: do not mount fonts cache on distros with quirks
* interfaces/builtin/desktop: do not mount fonts cache on distros with quirks

There is a problem with compatibility of fontconfig cache files generated across
different versions of the fontconfig library. The cache files appear to be using
the same format, but when loaded by an older version of the library, they cause
instability, bad font rendering or plain segfaults.

The following versions of packages present on specific distros are known to
generate incompatible cache files:

- Arch: fontconfig 2:2.13.91+24+g75eadca-2
- Fedora 32: fontconfig-2.13.92-9.fc32.x86_64

More information is provided in the following post:
https://forum.snapcraft.io/t/snapped-app-not-loading-fonts-on-fedora-and-arch/12484/32
A bisect identified this commit where the problem first occurred:
https://gitlab.freedesktop.org/fontconfig/fontconfig/commit/0f9bbbcf8f6f8264efb0a2ded4d8d05f3b10f7a4

Until we work out a better solution, disable the fonts cache on distros that are
known to cause issues. This comes at the potential cost of slightly longer
application startup times, but a working app is better than a segfaulting one.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>

* interfaces/builtin/desktop: mention the fonts bug report in comments

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>

* tests/main/interfaces-desktop-host-fonts: update fonts cache test

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>