snap/squashfs: also symlink snap Install with uc20 seed snap dir layout (#8711)
This ensures that during install/recover mode we use symlinks in /var/lib/snapd/snaps to /var/lib/snapd/seed/systems//snaps, and that makebootable will still copy the full snap file to /run/mnt/ubuntu-data/... such that the initramfs and run mode see a normal snap file. Note that we specifically don't want the initramfs to use symlinks from ubuntu-data to ubuntu-seed because this is not trusted, this is enforced by passing a new InstallOptions to Container.Install with a MustNostCrossDevices flag.

This was prompted by https://bugs.launchpad.net/snapd/+bug/1878647 because there, the tmpfs we mount for ubuntu-data/writable is only 205M, and all the kernel and base snaps are over 205M and so the install runs out of space to copy the snap files onto the tmpfs and eventually snapd tries to uninstall itself during first-boot.

This fixes https://bugs.launchpad.net/snapd/+bug/1878647


* boot/makebootable: if kernel,base snaps are symlinks, follow symlink

This ensures that if we switch setup during install/recover mode to use symlinks
in /var/lib/snapd/snaps to /var/lib/snapd/seed/systems/<label>/snaps, that
makebootable will still copy the full snap file to /run/mnt/ubuntu-data/...
such that the initramfs and run mode see a normal snap file.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* snap/squashfs: fix creating symlinks when snap is in uc20 seed system snaps dir

During uc20 install and recover modes, we will have a tmpfs as our writable
partition and as such we don't really want or need to make a full copy of the
snap files into /var/lib/snapd/snaps, and can make do with a symlink instead.

However, during uc20 install and recover modes the snap files from the seed dir
will be of the form /var/lib/snapd/seed/systems<label>/snaps/<name>.snap and so
the previous logic to install it was wrong, as the <name>.snap would not be a
direct child of /var/lib/snapd/seed, and instead needs to check that it is just
prefixed with /var/lib/snapd/seed somewhere.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* snap/many: support InstallOptions + MustCopy for Install()

This will allow us from snapstate to decide how a snap file should be installed,
if it should be symlinked or if it should be always copied, etc. Eventually we
should also introduce logic to snapstate to decide if it should be linked to
remove the overlayfs specific logic from here where it is kind of awkward to
see.

Also add tests for the squashfs and snapstate implementations.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* overlord/snapstate: always copy files when seeding in uc20 run mode

This prevents us from inadventently installing symlinks from ubuntu-data to
ubuntu-seed, which would be insecure in the full disk encryption case because
an attacker could modify ubuntu-seed and after the disk is unlocked on
ubuntu-data and the system starts running, the snap from ubuntu-seed would be
mounted.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* snap: rename MustCopy to MustNotCrossDevices for InstallOptions

This option is really about security and ensuring that when we install a snap
that installation does not cross device/trust boundaries, i.e. on UC20 from
ubuntu-data (trusted) to ubuntu-seed (untrusted), so rename the option to better
reflect this.

Also, refactor the implementation to allow hard-links if MustNotCrossDevices is
true, as by definition hard-links cannot cross devices/filesystems, and really
should only limit symlinks.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* snap/snapdir: un-support MustNotCrossDevices in Install() for now

It's unclear what we want this implementation to look like, so for now just drop
it. `snap try` may be less secure on UC20 as a result, but that shouldn't be
happening on production devices anyways.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* tests/core/basic20: also test that no symlinks were created during seeding

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* tests/core/seed-base-symlinks: disable for UC20 permanently, adjust comment

We don't have symlinks from ubuntu-data /var/lib/snapd/snaps to ubuntu-seed at
/var/lib/snapd/seed by design on UC20 run mode.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* tests/core: fix spread tests adjustments

* squashfs: drop TODO:UC20 more likely to confuse than get ever addressed

Co-authored-by: Samuele Pedroni <pedronis@lucediurna.net>