tests: port xdg-open to session-tool
There are the usual suspects again, session-tool and retry-tool.
I dropped -f from umount as that option was a little bit misused.
It's a bind mount so we should have used --lazy, --force is for stale
NFS mounts. As in an earlier change to xdg-open-compat the restore and
prepare sections in the YAML are backwards to minimize the delta.

For context, this is almost the last test that leaks dbus-daemon, just
three more left.

Amazon Linux 2 and CentOS 7 do not support DBus session bus so that
they are left out.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>