dbusutil: move all D-Bus helpers and D-Bus test helpers
This commit is squashed history taken out of the refresh-app-awareness-v2
branch. It aggregated most of the new D-Bus helper logic in a pair of
packages. The dbusutil package contains non-silly logic to connect
to the session and system bus, as well as public mocking helpers.
The dbustest sub-package contains two helpers that provide dbus.Conn
for testing interaction and for testing bus access.

Note that I chose to leave behind one file in testutils, as it contains
existing test code used by user agent code. I can do that in another pass
to isolate it from this move.

This code is used in feature/refresh-app-awareness-v2, for both the
existing code in "snap run" and the new code in sandbox/cgroup package.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>