cmd/snap-confine/snap-device-helper: support nvme block devices too
We need to be a bit tricky here because some device paths that match
the nvme pattern are block devices while others are char devices. The
block devices are more deeply nested and show up like:

/devices/pci0000:00/0000:00:1d.0/0000:04:00.0/nvme/nvme0/nvme0n1

while the char devices show up like:

/devices/pci0000:00/0000:00:1d.0/0000:04:00.0/nvme/nvme0

so the easy way to do this is to strip off everything before "nvme/" in
the path, and then see if there are any "/" characters left, if there
are then we have a deeply nested path which is a block device, otherwise
it is a char device.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>