many: change SetRunKernelImageEnabledKernel to SetEnabledKernel
This will be useful in a future commit where we add SetEnabledKernel to the main
MockBootloader implementation, and then some tests in boot and other places can
use the same function with different implementations depending on if the
underlying bootloader.Bootloader is implemented with
ExtractedRunKernelImageBootloader or with just the
EnvRefExtractedKernelBootloader implementation that does not implement the
ExtractedRunKernelImageBootloader interface like i.e. u-boot.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>