sandbox/cgroup: add tracking helpers
This patch adds two related tracking helpers:
 - CreateTransientScope()
 - ConfirmSystemdServiceTracking()

Those two functions are complementary, the first ought to be used for
all hooks and the subset of applications that are not services.
The second function confirms that the current process is already
tracked as a part of a *snap specific* service.

Both functions can return ErrCannotTrackProcess in the case where
either tracking cannot be established or when the service cannot
be matched to a security tag.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>