bootloader/bootloadertest: add Mock20BootloaderSetup
This interface is necessary so we can avoid awkwardness in UC20 boot tests that
need to do minimal setup with the bootloader, just setting the kernel and
try-kernel, but for different bootloader implementations, namely
ExtractedRunKernelImageBootloader and EnvRefExtractedKernelBootloader.

This this in place we can expand the coverage of the initramfs tests in the boot
package to cover the u-boot EnvRefExtractedKernelBootloader implementation as
well.

Also fix a typo.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>