boot/initramfs_test.go: expand to MockUC20EnvRefExtractedKernelRunBootenv
This expands the set of tests we run for the InitramfsRunModeChooseSnapsToMount
function to also test the case where the bootloader is an
EnvRefExtractedKernelBootloader as well as an ExtractedRunKernelImageBootloader.

This requires us to be slightly more explicit in a few of the test cases,
because the implementation of SetEnabled{,Try}Kernel will set boot vars that we
want to also change separately to simulate weird situations the code is robust
against.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>