tests: autoremove after removing lxd in preseed-lxd test
This test was leaking the lxcfs mount point that caused failure of
mount-ns:inherit test, if scheduled accordingly. The root cause is that
lxd Debian package depends on a number of dependencies, including lxcfs,
which keeps a mount point active at all times.

To prevent this from happening again, add an invariant that detects this
case.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>