tests: port snap-handle-link test to tests.session
This test used DBus activation and leaves another dbus-daemon around. It
was picked up by the invariant detector.

This test is interesting because it makes it obvious why we set DISPLAY
to a fake value. I may circle back to earlier tests that have a similar
property and handle that better there.

In addition, xdg-open on the host can now correctly handle desktop files
on 16.04 so I made the test not skip that part.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>