tests/lib/prepare: increase the size of the uc16/uc18 partitions
Frequently, while locally developing things with random files even just like
compiled binaries of the various executables in the project, preparing a UC16 or
UC18 spread run will fail because the image built by ubuntu-image will have a
writable partition that is too small for these extra files.

It's also not feasible to just add these files to the gitignore or spread
tracking because these files could be named anything as spread picks up anything
that's not specifically excluded, so we would have to keep updating the spread
exclude list with all sorts of random, developer specific files that shouldn't
be included in the image.

To safeguard against this, we can just make the partition a bit bigger which
will solve almost all cases and make development experience a bit easier for
folks.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>