tests: use the "jq" snap from the edge channel
We are currently "maintaining" the jq snap because we need it on
core devices. However the current stable version is still at 1.5
which is slightly embarrassing. We have 1.6 in edge so this commit
switches to use jq from edge to ensure we have no regresions. We
will keep using edge to ensure that we always test with the latest
jq.