data/sudo: drop a failed sudo secure_path workaround
The sudo secure_path setting resets the PATH to some predefined value for
commands executed under sudo. We have tried to workaround
https://bugzilla.redhat.com/show_bug.cgi?id=1691996 by trying to extend the
secure_path in a drop in conf files. This approach does not work for 2 reasons:
 - the file is incorrectly named
 - secure_path is a string and += append only works on lists (eg. env_keep)

Since there is no clear way to fix the problem other than talking with
distributions, drop the workaround. We can always revert the patch when needed.

Fixes: https://bugs.launchpad.net/snapd/+bug/1882215

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>