snapstate: fix autorefresh from classic->strict
We got a bugreport that going from classic->strict in a refresh
does not work. This commit fixes this issue.

Back in 864e0e861d1 code was added in validateFlagsForInfo() to
check that it is not possible to do:
```
$ snap install --classic test-snapd-tools
```
i.e. this resulted in an error with this check.

However in d983429b29e this was changed so that it is now possible
to do such an install. If this happens the --classic option is
simply ignored.

This commit now changes the code from d983429b29e to also apply
to refreshes, not just installs.  This fixes refreshes that go
from classic->strict. This unblocks multipass (LP: #1883538).