osutil/disks: use ID_FS_LABEL_ENC and encode user provided label to compare
Also update the partitions map to fsLabelToPartUUID, as that is closer to what
it is, a mapping between filesystem labels to partition UUIDs.

Also add some TODOs about combining with other code we already have that is
somewhat similar.

Finally, always use the last partition with a given filesystem label to save in
the map, this matches what udev does so is a sensible choice.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>