tests/lib/nested.sh: eliminate race after refreshing core snap
After we reboot for a core snap refresh, there is a short window after SSH is
available before the refresh change is fully done. In order for code to operate
more reliably, we can wait for that change to be done after we come back from
the reboot. This fixes the tests/nested/core/core-snap-refresh-on-core which
has been observed to frequently fail from this for example.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>