osutil/disks: adjust error message handling
We were missing errors that are not os.IsNotExist errors, so now we just handle
all errors with the same message. The previous attempt was an attempt to make a
smaller / more readable error message, but at this point it's probably okay to
leave it as-is.

Also add some more detail to another error message as if it fails it might be
indicative of a change in upstream udev code which we would want to know about
from the error message.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>