many: add new "daemon-startup: inhibit" option
This commit adds a new options to inhibit startup of daemons. The
default of snapd is that all daemons are started once you install
the snap. However that is sometimes not desired and our users
need to write fragile wrapper scripts to workaround the missing
native support.

With the new `daemon-startup: inhibit` the service is not started
on install and can be enabled via snapctl from a hook or via a
management agent.