tests/core/uc20-recovery: add hack for external back to copy gopath
With the external backend, we do not have the special snapd snap like we do on
the google backend which copies gopath data from ubuntu-seed to the current
ubuntu-data (which in recover mode is a tmpfs). Instead, we use a small hack,
abusing the ~/.ssh/rc file which is run on the first ssh connection from spread,
and will copy the required files over.

We need to remove this ~/.ssh/rc file before going back to run mode however, as
otherwise spread fails to re-connect with:

ssh: unexpected packet in response to channel open: <nil>

Unclear what makes that happen, as a normal ssh connection from i.e. a terminal
works fine with this ~/.ssh/rc file in place.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>