overlord: mock timings.DurationThreshold in TestNewWithGoodState
The TestNewWithGoodState test checks that reading/writing the
state works correctly. However when the loading is longer than
timings.DurationThreshold which is 5ms right now then the test
will fail because the state contains more data than the test
expects. This can happen on really slow architectures like
the emulated risc-v.

To fix this this commit just increase timings.DurationThreshold
during the test.