sandbox/cgroup: fail tests if sequence if is violated
Tests heavily rely on mocking and return errors from some mocked
functions to indicate that the function fails or if the real
implementation is undesired even though the test is not expected to use
it. In the latter case, make sure that test fails by calling c.Error, so
that even if the returned error object is ignored, the test is reliably
marked as failed.

Thanks to Ian for the suggestion.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>