osutil: detect autofs mounted in /home
For a while, we had reports about autofs being incompatible with snapd.
We just got extra information about how an autofs setup looks like.

There are two participating components, autofs uses an automount mount
point, which is a kernel feature that mounts a filesystem whenever a
path is accessed in a normal way, and an userspace service that responds
to those mount requets and performs the actual mount operation on when
asked by the kernel.

We can expand the existing NFS detection to trigger on autofs and treat
it the exact same way. We could attempt to parse autofs configuration as
well but in practice it seems to be used for network filesystems so I
don't think this is required.

Ref: https://www.kernel.org/doc/html/latest/filesystems/autofs.html
Ref: https://www.kernel.org/doc/html/latest/filesystems/automount-support.html
Fixes: https://bugs.launchpad.net/snapd/+bug/1784774
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>