tests: fix leaked dbus-daemon in selinux-clean
There's one more D-Bus daemon that was leaking from the selinux-clean
test, due to that test's usage of the desktop interface. This failure
was masked by the unstable nature of the CentOS 8 image where we ignore
known errors and the invariant violation fell through the cracks.

Thanks to Sergio for spotting.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>