cmd/snap: track started apps and hooks
When refresh-app-awareness is enabled, all non-service programs are
automatically tracked by asking systemd to move the launcher process
to a transient scope.

There are associated spread tests for the tracking alone. Note that
refresh-app-awareness is not yet using this additional information.
This will come in subsequent patches.

To aid testing, tests.session now forwards SNAPD_DEBUG and
SNAP_CONFINE_DEBUG across the session boundary, so that observing debug
output is easy.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>