snap/validate.go: disallow snap layouts with new top-level directories
We previously wouldn't fail on verification for snaps that used new top-level
directories, and would fail at runtime, which produces a very unhelpful message.

Also adjust some test yamls in interfaces/mount tests, which are not valid as
they are for new top-level directories in /.

Finally, add an additional unit test for the test case with LP #1831010 that is
already covered by spread tests, but not yet by unit tests.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>