systemd/escape: fix issues with "" and "\t" handling
"" is treated as "/", however filepath.Clean("") gives "." which then gets
escaped and is incorrect.

"\t" needs the prefixing 0 in the \x escaping to match systemd-escape(1)
exactly, the easy way to have Go do this automatically is to use %x format spec
with a byte list.

Add many more unit tests here and stop importing the package with "." as it is
slightly confusing.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>