systemd/escape: add additional invalid unit test case, fix comment
The invalid case here causes systemd-escape(1) to fail thusly:

$ systemd-escape -p /a/../b/
Failed to escape string: Invalid argument
$ echo $?
1

however since it's not documented how or why systemd-escape fails like this,
rather than rely on internal implementation details of why systemd-escape fails
here, we just accept our current implementation while testing that we are
consistent with behavior for this type of path. If systemd-escape(1) official
docs ever explain why this is considered invalid then we can try to deny this
too, but until then leaving it as is probably fine.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>