tests/cmd/snap-bootstrap/initramfs-mounts: add test case for empty recovery_mode
This is in install mode, so we should have snapd_recovery_mode=install here.

The empty string here is still interpreted as install mode so this is a pure
test change. Eventually we should consider dropping support for
snapd_recovery_mode="" meaning install mode, but for now just test both cases.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>