tests: adjust xdg-open after launcher changes
Commit d8a2b003ad911b063f318f93076fb4744ed94ccd removed the code path
where userd launcher resolved the pid of the calling process to a snap
name - something that currently only works with cgroup v1, without
properly adjusting the test. Without this code the launcher is no longer
sensitive to cgroup v1 vs v2 and the expectation in the test is no
longer upheld.

This change was merged in https://github.com/snapcore/snapd/pull/9022
but apparently, since it was a cherry pick from a security release, the
review was hasty and the fact this broke master was unnoticed.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>