tests: check for pids correctly.
Using grep for PIDs is easy to do incorrectly, as we've found out the
hard way when a test searching for pid 2499 found process 24998. Rewrite
that part of the test to use kill with signal zero as a much safer and
less magic way of doing the same thing.

As a drive-by fix a typo in the description and remove useless sleep
loops. One sleep of moderate duration is sufficient.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>