.github/workflows/snap-build.yaml: use globbing w/ v2 upload action
The name here actually is the name of a zip file that gets downloaded,
not the name of the file that is uploaded, so we should name it more
generic, and we can get all snaps that were built if someday we want to
build multiple snaps here (i.e. the core snap as well).

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>