overlord: use new tracking cgroup for refresh app awareness
With all of the bits now in place we can switch the overlord to
measure running apps using the new tracking cgroup, not the pids
cgroup.

The pids code is left intact but not used, it will be removed
in a subsequent commit.

The refresh-app-awareness spread test is extended to test all systems
except 14.04 (where the experimental feature no longer works). The
LXD test is extended to ensure that refresh app awareness can be enabled
both inside and outside of containers, and more importantly, can be
configured in all the possible combinations without regressions.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>