cmd/snap/debug/seeding: display "-" for times we can't calculate
Making the times pointers means we won't emit empty time.Time server side and
thus we can more easily know when we can't make a duration calculation, i.e.
seed time of 2020, but no seed-start-time does not mean that seeding took 2
millenia. In cases like those we just display "-" instead.

Note that this isn't realistic until the daemon also outputs *time.Time in it's
JSON response but that can happen OOB from this PR.

Also round the times to the nearest millisecond, as we don't really have much
better precision and it's misleading to represent that.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>