secboot,cmd/snap-bootstrap: cross-check partitions before unlocking, mounting
Before we mount partitions, we should check / verify that the partition we want
to mount is coming from the same disk that we have already mounted partitions
from, or in the original case, that the partition comes from the same disk as we
booted the kernel from.

In the unlocking case in secboot, we should only try to unlock an encrypted
partition if it comes from the same disk that we mounted the first partition
from, i.e. ubuntu-boot in run mode and ubuntu-seed in recover mode.

Additionally, we should verify that the decrypted partition we got after
unlocking comes from the same encrypted partition that we requested it come
from.

Fixes: https://bugs.launchpad.net/snapd/+bug/1863886
Signed-off-by: Ian Johnson <ian.johnson@canonical.com>