osutil/disks: use a dedicated error to indicate a fs label wasn't found (#9080)
* osutil/disks: use xerrors to indicate a fs label wasn't found

It will be necessary in the secboot package to positively indicate that a FS
label was not found and that we didn't have some other error, as the first step
to unlocking is to check if there is an encrypted device. We currently identify
that by using /dev/disk/by-label/ubuntu-data-enc, but this is insecure and buggy
as it means if the boot disk was not encrypted (i.e. on a device with a TPM but
without FDE enabled/used), an attacker can attach a USB drive with a FS label of
ubuntu-data-enc (as opposed to the real boot disk with ubuntu-data label on it),
and now the system will fail to boot because the ubuntu-data-enc USB drive FS
label will not be unlockable. We will avoid this bug by checking if the disk
that we booted the kernel from (and mounted the first partition using) has a
partition with a FS label on it called ubuntu-data-enc, rather than any
partition with a FS label of ubuntu-data-enc.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* osutil/disks: use fmt string var instead of string directly for old gofmt

Old gofmt doesn't understand the new %w format specifier for error wrapping, but
we can avoid the problem by putting the format string into a variable.

Also rename FilesystemLabelNotFound to ErrFilesystemLabelNotFound and add a
doc-comment on it.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* osutil/disks: don't use xerrors.Errorf, just create error type struct

This presents a better error message and is still comparable using xerrors.As()
or just doing the old style of type assertions on the returned error.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* osutil/disks: naming is hard

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>