cmd/snap/debug/seeding: use unicode for proper yaml (#9090)
In order for the output to always be valid yaml, we need to escape the "-" we
output when we cannot calculate a time, as that is not valid yaml. If we have
unicode, then we can output a special unicode dash that looks nice like a
standard dash, but is yaml, and if we don't have unicode we can output just two
dashes instead.