boot/bootstate20: unify commit method impls, rm bootState20MarkSuccessful
Refactor the various structs in bootstate20.go to use a common bootStateUpdate
implementation with a single commit() method that is used for all scenarios.
This is necessary because we will soon need to reseal against the TPM when
we commit boot state changes, so having that happen in one place will greatly
simplify things.

This unification allows us to remove one of the structs,
bootState20MarkSuccessful entirely, and allows us to greatly reduce the
amount of state that the bootState20Kernel and bootState20Base structs
carry around.

Also fix a typo in the comment for revisions(), as isTrySnapError is not
exported anymore.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>