docker-support: also allow rix on files with privileged-containers
containerd by default sets up containers with pivot_root which makes
files in the container show up in arbitrary locations. c9a8a13fb73 added
a rule for '/** wl' so extend that to be '/** rwlix'. This doesn't
affect the security stance of the interface significantly (since the
prior 'w' would allow writing to a location covered by an existing ix
rule) and the privileged-containers variant of the policy grants device
ownership to the snap already.