many: remove usage and creation of hijacked pid cgroup
Snapd used to rely on a hijacked PID cgroup for application tracking.
This was never enabled by default, as it was behind a feature flag.
We've recently landed the new systemd-asisted tracking logic so this
is entirely useless now.

I've explicitly left out the changes that impact the workaround
for systemd raciness in snap-run. The workaround is no longer required
as now there is only a single writer, but I wanted to have a separate
review for that.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>