bootloader: add helper for creating a bootloader based on gadget
The bootloader.Find() expects bootloader configuration to be present in the root
directory. However, when we are installing a run mode system, the bootloader
config is not present until after InstallBootConfig() gets called, which happens
only after the filesystem has been created and populated. However, when using
the ContentObserver to track trusted assets, we must know what the bootloader
is.

Add a helper for creating a bootloader based on what is carried by the gadget.
Move the basic checks for config file existence to Find().

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>