runinhibit: open the lock file in read-only mode in IsLocked
This allows unprivileged "snap run" process to safely examine a lock it
doesn't otherwise have permission to create. Note that this allows snap
run to open an existing lock file, read the hint but not write a hint.
Both snapd and snap run can flock a file (a file opened for reading can
still be flock-ed), which is relevant for synchronizing access to the
hint stored therein.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>