cmd/snap-confine: explain how fds move across functions
The code in snap-confine relies heavily on the C extension providing
cleanup attributes. We use cleanup attributes for safety, as they are
invoked in all return paths from the function using them. Sometimes we
need to provide an error path that closes an open descriptor and
sometimes we need to return a descriptor (or set of descriptors) in a
successful case. This patch expands the documentation, and adds a few
more cleanup attributes, so that the flow is more understandable.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>