tests: find -ignore_readdir_race when scanning cgroups
I noticed a random failure of the test
snap-confine-undesired-mode-cgroup that can be explained by a concurrent
modification of the /sys/fs/cgroup directory (most likely by systemd).
The error looked like this, note that find(1) failed to find a cgroup
directory related to kbdsettings.service - a suse-specific, short-lived
service that sets up keyboard configuration such as delay and repeat
rate.

    Error executing google:opensuse-tumbleweed-64:tests/main/snap-confine-undesired-mode-group:
    -----
    + su test -c 'snap run test-snapd-app.sh -c /bin/true'
    + for dname in /run/snapd /sys/fs/cgroup /tmp/snap.*
    + find /run/snapd -user test '!' -path '*/user@12345.service*' '!' -path '*/user-12345.slice*'
    + find /run/snapd -group test '!' -path '*/user@12345.service*' '!' -path '*/user-12345.slice*'
    + find /run/snapd '!' -type s '!' -type l '!' -name cgroup.event_control '!' -path '/tmp/snap.*/tmp' -perm /o+w
    + for dname in /run/snapd /sys/fs/cgroup /tmp/snap.*
    + find /sys/fs/cgroup -user test '!' -path '*/user@12345.service*' '!' -path '*/user-12345.slice*'
    find: ‘/sys/fs/cgroup/unified/system.slice/kbdsettings.service’: No such file or directory

This can happen when find(1) uses readdir reports a file name but the
file is removed before find can stat it. For this purpose, find sports
the option -ignore_readdir_race which makes that error silently ignored.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>