tests: use "set -ex" in prep-snapd-in-lxd.sh
We got some confusing test failures in the lxd test. It turns
out the "purge" in the nested container fail. But we did not
notice because the script `prep-snapd-in-lxd.sh` has
```
 #!/bin/sh -ex
...
```
but we run the script via:
```
lxd.lxc exec $cont_name -- /bin/bash /root/prep-snapd-in-lxd.sh
```
so the `-e` is never set.