packaging: umount /snap on purge for good measure
We observed a bunch of failures inside the lxd snap test. It
looks like there is a mountpoint on top of /snap that keeps
the directory busy. By umounting /snap this works. This is
very much a proof-of-concept at this point, we should probably
look at making this umount a bit more clever and less blunt.