boot/initramfs_test.go: reset boot vars on the bootloader for each iteration
This was missed in the initial implementation, but doesn't actually affect any
test results (probably by accident due to the order of the test cases not
depending on boot vars being after the ones depending on boot vars).

Missing this does make tests fail with changes I will be proposing separately,
so it will make for simpler review of that behavior change to see there is no
test infrastructure change for that, only test case behavior change.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>