tests/nested/core20/kernel-failover: add test for failed refresh of uc20 kernel
This exercises both the snapd logic for falling back properly and cleaning up
after a failed refresh, as well as pc-gadget grub.conf logic that rollbacks are
handled properly.

Eventually with resealing there should be additional things we check for here,
but for now it's sufficient to show that the kernel got rolled back successfully
and that we don't leave any intermediate state behind about the attempted
refresh.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>