cgroup,snap: track hooks on system bus only
I left a small TODO that became relevant just now. Hooks were using the
same uid-based tracking decision. For root user we would try the session
bus first and only later fall back to system bus.

For hooks it makes no sense to track them on the session bus, they are
not a property of the root session. They are a system-wide property.

Introduce TrackingOptions and allow CreateTransientScopeForTracking to
optionally use it to disable attempts to use the session bus. Use the
new capability to ensure that hooks do not use the session bus for
tracking.

Existing unit tests are extended and adjusted to observe the new behavior.
The spread test is extended to measure tracking of hooks and system-wide
services.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>